Allow running with default config from repo
Running as an unprivileged user is no problem, but the config
points to /var/lib/rex, which is usually a privileged directory.
Point to the current directory for all Minion/Mojolicious storage.

Don't get the rex and rexify executables from @krimdomu's hardcoded
home directory, but use the home directory of the current user.